requirements.txt
torch==2.2.2
torchvision==0.17.2
diffusers==0.31.0
peft==0.13.2
huggingface-hub==0.26.2
numpy==1.26.4
matplotlib==3.9.2
jupyter==1.1.1
click==8.1.7
frozenlist==1.5.0
insightface==0.7.3
onnx==1.17.0
pillow==10.2.0
facenet-pytorch==2.6.0
image-reward==1.5
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
